:sparkles: :bug: add unittests, rename package, fix bugs
