+++ tests/calls.json
@@ -1,0 +1,7 @@
+{"READY":
+	{"v":6,"user_settings":{},"user":{"verified":true,"username":"Piscord Unittest","mfa_enabled":false,"id":"715274051374940202","flags":0,"email":null,"discriminator":"9362","bot":true,"avatar":null},"session_id":"2257dd02df9e73cce0b0148d23ed9193","relationships":[],"private_channels":[],"presences":[],"guilds":[{"unavailable":true,"id":"715273516555174009"}],"application":{"id":"715274051374940202","flags":0},"_trace":["[\"gateway-prd-main-pg95\",{\"micros\":107058,\"calls\":[\"discord-sessions-prd-1-19\",{\"micros\":103617,\"calls\":[\"start_session\",{\"micros\":99454,\"calls\":[\"api-prd-main-us-east1-bf00\",{\"micros\":96085,\"calls\":[\"get_user\",{\"micros\":3401},\"add_authorized_ip\",{\"micros\":5509},\"get_guilds\",{\"micros\":44730},\"coros_wait\",{\"micros\":1}]}]},\"guilds_connect\",{\"micros\":2,\"calls\":[]},\"presence_connect\",{\"micros\":2039,\"calls\":[]}]}]}]"]},
+"GUILD_CREATE":
+	{"premium_tier":0,"large":false,"system_channel_flags":0,"region":"us-east","members":[{"user":{"username":"Astremy","public_flags":128,"id":"263331548542009348","discriminator":"2316","avatar":"a_12b9c81ae06e7d9baefbcd5d835b3f5d"},"roles":[],"mute":false,"joined_at":"2020-05-27T18:41:35.494000+00:00","hoisted_role":null,"deaf":false},{"user":{"username":"Piscord Unittest","id":"715274051374940202","discriminator":"9362","bot":true,"avatar":null},"roles":["715275047325007933"],"premium_since":null,"nick":null,"mute":false,"joined_at":"2020-05-27T18:47:40.463474+00:00","hoisted_role":null,"deaf":false}],"banner":null,"premium_subscription_count":0,"channels":[{"type":4,"position":0,"permission_overwrites":[],"name":"Salons textuels","id":"715273516555174010"},{"type":4,"position":0,"permission_overwrites":[],"name":"Salons vocaux","id":"715273516555174011"},{"type":0,"topic":null,"rate_limit_per_user":0,"position":0,"permission_overwrites":[],"parent_id":"715273516555174010","name":"général","last_message_id":null,"id":"715273516555174012"},{"user_limit":0,"type":2,"position":0,"permission_overwrites":[],"parent_id":"715273516555174011","name":"Général","id":"715273516555174013","bitrate":64000}],"icon":null,"joined_at":"2020-05-27T18:47:40.463474+00:00","features":[],"emojis":[],"max_video_channel_users":25,"verification_level":0,"lazy":true,"member_count":2,"mfa_level":0,"default_message_notifications":0,"public_updates_channel_id":null,"unavailable":false,"name":"Piscord_Unittest","guild_hashes":{"version":1,"roles":{"omitted":false,"hash":"HrmQrlokDgs"},"metadata":{"omitted":false,"hash":"nYGweUJem5g"},"channels":{"omitted":false,"hash":"J1n4YoPDApM"}},"discovery_splash":null,"system_channel_id":null,"afk_channel_id":null,"application_id":null,"preferred_locale":"en-US","vanity_url_code":null,"owner_id":"263331548542009348","presences":[{"user":{"id":"263331548542009348"},"status":"online","game":{"type":4,"name":"Custom Status","id":"custom","emoji":{"name":"GWofxLove","id":"329700139520360448","animated":false},"created_at":1590581106955},"client_status":{"desktop":"online"},"activities":[{"type":4,"name":"Custom Status","id":"custom","emoji":{"name":"GWofxLove","id":"329700139520360448","animated":false},"created_at":1590581106955},{"type":0,"timestamps":{"start":1590070067318},"name":"Sublime Text","id":"928e6f8f543ec997","created_at":1590581106991}]},{"user":{"id":"715274051374940202"},"status":"online","game":null,"client_status":{},"activities":[]}],"splash":null,"rules_channel_id":null,"roles":[{"position":0,"permissions":104320577,"name":"@everyone","mentionable":false,"managed":false,"id":"715273516555174009","hoist":false,"color":0},{"position":1,"permissions":8,"name":"Piscord Unittest","mentionable":false,"managed":true,"id":"715275047325007933","hoist":false,"color":0}],"description":null,"explicit_content_filter":0,"voice_states":[],"afk_timeout":300,"id":"715273516555174009"},
+"MESSAGE_CREATE":
+	{"type":0,"tts":false,"timestamp":"2020-05-28T10:37:37.693000+00:00","pinned":false,"nonce":"715514130617860096","mentions":[],"mention_roles":[],"mention_everyone":false,"member":{"roles":[],"mute":false,"joined_at":"2020-05-27T18:41:35.494000+00:00","hoisted_role":null,"deaf":false},"id":"715514111089442847","flags":0,"embeds":[],"edited_timestamp":null,"content":"test","channel_id":"715273516555174012","author":{"username":"Astremy","public_flags":128,"id":"263331548542009348","discriminator":"2316","avatar":"a_12b9c81ae06e7d9baefbcd5d835b3f5d"},"attachments":[],"guild_id":"715273516555174009"}
+}
