+++ tests/imports.py
@@ -1,0 +1,6 @@
+import os
+import sys
+
+sys.path.insert(0, os.path.abspath('..'))
+
+from piscord import *
